Fixed statistic type

+++ apps/prairielearn/src/migrations/20240325150316_assessment_questions__add.sql
@@ -1,2 +1,2 @@
 ALTER TABLE assessment_questions
-ADD COLUMN median_question_score NUMERIC;
+ADD COLUMN median_question_score DOUBLE PRECISION;
